Reorganize go code and add debug capabilities
- Setting DEBUG=true as env variable will enable debug messages when running the go code
- Move registry code to its own file
- Move gofalcon code to its own file
- Update WriteToCollection function to use an interface instead of ImageList
