+++ README.md
@@ -17,7 +17,13 @@ The pages will now be available in the Falcon console (after refresh) under the 
 
 Foundry functions don't have a "dev mode" so we test locally and then deploy:
 
-1. Load a `FALCON_CLIENT_ID`, `FALCON_CLIENT_SECRET`, and `FALCON_CLOUD` into your environment
+1. Export the following variables into your environment:
+   -  `FALCON_CLIENT_ID`
+   - `FALCON_CLIENT_SECRET`
+   - `FALCON_CLOUD`
+   
+   If you would like to enable debug logs, export `DEBUG=true` as well.
+
 1. In `functions/syncimages`:
 
    1. `go run main.go`
